chore: remove impl layer and move all business logics to EventService class
